% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rparallelcuda-package.R
\docType{package}
\name{rcudarmadillo-package}
\alias{rcudarmadillo}
\alias{rcudarmadillo-package}
\title{rcudarmadillo: Package to test CUDA with RcppArmadillo}
\description{
Benchmark of armmadillo with CUDA.
}
\keyword{internal}
