% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{solve_problem}
\alias{solve_problem}
\title{Matrix solver}
\usage{
solve_problem(rows, cols, iterations)
}
\arguments{
\item{rows}{`numeric` matrix rows}

\item{cols}{`numeric` matrix columns}

\item{iterations}{`numeric` number of tries}
}
\description{
Matrix solver
}
