% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solveMatrix.R
\name{solveMatrix}
\alias{solveMatrix}
\title{An example to solve a system}
\usage{
solveMatrix(A = NA, B = NA, size = 200)
}
\arguments{
\item{A}{`matrix` `numeric` input matrix}

\item{B}{`matrix` `numeric` input matrix}

\item{size}{`numeric` size of square matrix}
}
\description{
An example to solve a system
}
