% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solveMatrix.R
\name{solveMatrix}
\alias{solveMatrix}
\title{An example to solve a system}
\usage{
solveMatrix(rows = 300, cols = 200, iterations = 20)
}
\arguments{
\item{rows}{`numeric` number of rows}

\item{cols}{`numeric` number of columns}

\item{iterations}{`numeric` number tries}
}
\description{
An example to solve a system
}
